refactor button for pagination
