+++ src/components/List/interface.ts
@@ -1,10 +1,6 @@
 import { ReactNode } from "react";
 
-export interface IListTranslations {
-  fetchMore: string;
-}
-
-export interface IListContainerProps<ListItem> {
+export interface IListProps<ListItem> {
   list: ListItem[];
   className?: string;
   fetchMoreClassName?: string;
@@ -12,7 +8,4 @@ export interface IListContainerProps<ListItem> {
   fetchMore?: () => void;
   shouldFetchMore?: boolean;
   loading: boolean;
-}
-export interface IListProps<ListItem> extends IListContainerProps<ListItem> {
-  translations: IListTranslations;
 }
