+++ src/components/List/styles.module.scss
@@ -5,14 +5,3 @@
   flex-direction: column;
   justify-content: space-between;
 }
-
-.fetchMore {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  margin: $spacing-m 0 0;
-}
-
-.fetchMoreButton {
-  margin: 0 auto;
-}
