+++ src/components/PaginationButton/index.tsx
@@ -1,0 +1,1 @@
+export { PaginationButtonContainer as PaginationButton } from "./container";
