+++ src/components/PaginationButton/styles.module.scss
@@ -1,0 +1,5 @@
+@import "$styles/_variables";
+
+.fetchMoreButton {
+  margin: 0 auto;
+}
