+++ src/pages/Admin/components/ListPageContainer/List/ListContent/interface.ts
@@ -1,18 +1,10 @@
 import { Listable, ListableReactNodes } from "../../interfaces";
 
-export interface IListContentContainerProps {
+export interface IListContentProps {
   items: Listable[];
   rowClassName: string;
   children: ListableReactNodes;
   fetchMore?: () => void;
   shouldFetchMore?: boolean;
   loading: boolean;
-}
-
-export interface IListContentTranslations {
-  fetchMore: string;
-}
-
-export interface IListContentProps extends IListContentContainerProps {
-  translations: IListContentTranslations;
 }
